+++ modules/SofaNonUniformFem/DynamicSparseGridTopologyModifier.cpp
@@ -98,7 +98,7 @@ void DynamicSparseGridTopologyModifier::renumberAttributes( const sofa::helper::
     helper::vector<BaseMeshTopology::HexaID>& iirg = *m_DynContainer->idxInRegularGrid.beginEdit();
 
     // Update the data
-    unsigned int nbElt = (unsigned int)iirg.size();
+    unsigned int nbElt = iirg.size();
     std::map< unsigned int, BaseMeshTopology::HexaID>& regularG2Topo = *m_DynContainer->idInRegularGrid2IndexInTopo.beginEdit();
     for ( sofa::helper::vector<unsigned int>::const_iterator it = hexahedra.begin(); it != hexahedra.end(); ++it )
     {
